Fix pointer lock events not allowing to release Module instance
The commit contains rough fix to the problem where pointer lock events were created and never cleared from the memory.
Therefore Module instance was unable to be garbage collected and release memory properly, staying in the memory even after all references to it were lost.
This fix has a downside of pointer lock support being broken now in the application, so it's not suitable for the official emscripten repository.